NFC: move the MIFARE Ultralight scenes into its plugin
First protocol onto the extra_scenes mechanism from the previous commit. The
six Ultralight-only scenes - both dictionary attacks, the unlock menu, warning,
key input and password capture - move into nfc_mf_ultralight.fal. The app keeps
a thunk per scene under scenes/, so every existing scene id and every
search_and_switch_to_previous_scene call site still resolves exactly as before.

Not one line of their logic changed. The Ultralight C and AES attacks differ in
ways that matter - AES records success from AuthSuccess and handles ReadFailed,
C infers it from a full page read - and moving preserves both rather than
forcing a choice, which is why this is the right tool for them and merging was
not.

Mechanical adjustments only: handlers take NfcApp* and are renamed
mf_ultralight_scene_* inside the plugin, because nfc_scene.h still declares the
nfc_scene_* names with a void* signature for the thunks. Both attacks defined
the same DictAttackState enum and it is now defined once. One dead helper,
nfc_scene_mf_ul_c_dict_attack_update_view, went with them - it had no callers
and only compiled because it was non-static.

The dictionary attack view joins the app API table. That is a shared piece of
app UI driven by four protocols, not protocol logic, so it stays in the app and
plugins reach it through nine setters - the opposite of the mosgortrans case,
where 8.5 KB of single-consumer parsing was pinned for no reason.

Resident image: 60,903 -> 57,483 B (-3,420). nfc_mf_ultralight.fal grows
9,617 -> 13,749 B, paid only while an Ultralight card is loaded.
